chore: split tests and make ms
